+++ configs/file.ini
@@ -1,8 +1,8 @@
 [file]
-upload_path = uploads/
+upload_path = assets/upload
 
 [file.image]
-upload_path = uploads/images/
+upload_path = assets/upload/images/
 # MB
 max_size = 5
 allow_extends = .jpg,.jpeg,.png,.gif
