+++ internal/app/api/user.go
@@ -1,11 +1,11 @@
 package api
 
 import (
-	"github.com/DowneyL/now/models"
-	gresp "github.com/DowneyL/now/packages/http/response"
-	"github.com/DowneyL/now/packages/locales"
-	"github.com/DowneyL/now/packages/util"
-	"github.com/DowneyL/now/services"
+	"github.com/DowneyL/now/internal/models"
+	"github.com/DowneyL/now/internal/services"
+	gresp "github.com/DowneyL/now/pkg/http/response"
+	"github.com/DowneyL/now/pkg/locales"
+	"github.com/DowneyL/now/pkg/util"
 	"github.com/gin-gonic/gin"
 )
