--- routers/api.go
@@ -1,25 +1,0 @@
-package routers
-
-import (
-	apiController "github.com/DowneyL/now/controllers/api"
-	v1 "github.com/DowneyL/now/controllers/api/v1"
-	middleware "github.com/DowneyL/now/middlewares"
-	"github.com/gin-gonic/gin"
-)
-
-func api(r *gin.Engine) {
-	group := r.Group("/api")
-	{
-		group.GET("/database/migrate", apiController.Migrate)
-		group.POST("/user", apiController.Register)
-		group.POST("/user/login", apiController.Login)
-	}
-}
-
-func apiV1(r *gin.Engine) {
-	group := r.Group("/api/v1")
-	group.Use(middleware.Jwt())
-	{
-		group.POST("/user/email", v1.SetEmail)
-	}
-}
